#47 GameDispageのBoard,Column,Cellコンポーネントを新しく作成したboardディレクトリに配置
