+++ src/board/Cell.js
@@ -1,0 +1,18 @@
+import React from "react";
+import "./../pages/GameDisplayPage.css";
+
+const Cell = (props) => {
+  let color = "white";
+  if (props.value === "Player1") {
+    color = "red";
+  } else if (props.value === "Player2") {
+    color = "yellow";
+  }
+  return (
+    <div className="cell" data-x={props.x} data-y={props.y} onClick={props.onClick}>
+      <div className={color}></div>
+    </div>
+  );
+};
+
+export default Cell;
