+++ src/board/Column.js
@@ -1,0 +1,15 @@
+import React from "react";
+import "./../pages/GameDisplayPage.css";
+import Cell from "./../board/Cell.js";
+
+const Column = (props) => {
+  return (
+    <div className="flex-direction-column-reverse">
+      {props.column.map((y, i) => (
+        <Cell value={y} key={i} x={props.x} y={i} onClick={props.onClick} />
+      ))}
+    </div>
+  );
+};
+
+export default Column;
